pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "qwen3-encdec"
version = "0.1.0"
description = "Qwen3-based encoder-decoder model following T5Gemma 2 architecture"
readme = "CLAUDE.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Vince" }
]
keywords = ["transformers", "encoder-decoder", "qwen3", "nlp", "machine-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "transformers>=4.40.0",
    "torch>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
training = [
    "accelerate>=0.27.0",
    "datasets>=2.16.0",
    "wandb>=0.16.0",
    "tensorboard>=2.15.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
    "safetensors>=0.4.0",
    "UL2_5 @ git+https://github.com/pszemraj/UL2_5.git@v0.1.4",
]
optimizations = [
    "cut-cross-entropy>=24.11.0",
    "liger-kernel>=0.4.0",
]
evaluation = [
    "mteb>=1.0.0",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
]
all = [
    "qwen3-encdec[dev,training,optimizations,evaluation]",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/qwen3_encdec", "src/data", "src/training"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/qwen3_encdec"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
